% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{citT_competitions}
\alias{citT_competitions}
\title{citT competition data}
\format{
## `citT_competitions`
A data frame with 6 rows and 7 columns:
\describe{
  \item{strain1}{Designation for strain without rnk-citT module}
  \item{strain2}{Designation for strain with rnk-citT module}
  \item{other columns}{Competition data}
  ...
}
}
\source{
<https://doi.org/10.1371/journal.pgen.1007348>
}
\usage{
citT_competitions
}
\description{
Data showing the fitness effects of adding the rnk-citT module to the
genomes of Escherichia coli strains from the Long-Term Evolution Experiment.
}
\details{
Reference

Leon, D., D’Alton, S., Quandt, E.M., ^Barrick, J.E. (2018) Innovation in an 
E. coli evolution experiment is contingent on maintaining adaptive potential 
until competition subsides. PLoS Genet. 14:e1007348.
}
\keyword{datasets}
