% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{competition_template}
\alias{competition_template}
\title{Competition assay template}
\format{
\subsection{\code{competition_template}}{

A data frame with 12 rows and 11 columns:
\describe{
\item{competitor1}{the name of competitor 1 (e.g., strain or sample designation). Example: REL606}
\item{competitor2}{the name of competitor 2 (e.g., strain or sample designation). Example: REL8604}
\item{competitor1_initial_count}{competitor1_initial_count - the counts of competitor1 cells when the two competitors are mixed (e.g., colony counts). Example: 243}
\item{competitor2_initial_count}{competitor2_initial_count - the counts of competitor2 cells when the two competitors are mixed (e.g., colony counts). Example: 128}
\item{competitor1_final_count}{the counts of competitor1 cells after the competitors are co-cultured (e.g., colony counts). Example: 198}
\item{competitor2_final_count}{the counts of competitor 2 cells after the competitors are co-cultured (e.g., colony counts). Example: 271}
\item{transfer_dilution}{the overall dilution factor for all passages of the cells in co-culture. Example: 100  (If 2 x 50 µl of saturated cultures of each competitor were mixed into 9.9 mL of fresh medium and then the culture grew to saturation.)}
\item{initial_dilution}{the dilution factor for the initial sample that was plated. Example: 10000  (If 2 x 50 µl of saturated cultures of each competitor were mixed into 9.9 mL of fresh medium, then 100 µl of this was diluted into 9.9 ml of sterile saline.)}
\item{initial_volume}{volume of the initial dilution that was analyzed for counting each competitor.  The units on initial_volume and final_volume must be the same. Example: 80  (If 80 µL were plated.}
\item{final_dilution}{the dilution factor for the initial sample that was plated. Example: 10,000 if 100 µl of final saturated cultures was diluted into 9.9 mL of sterile saline, then 100 µl of this was diluted into 9.9 ml of sterile saline. Example: 10000  (If 100 µl of final saturated cultures was diluted into 9.9 mL of sterile saline, then 100 µl of this was diluted into 9.9 ml of sterile saline.)}
\item{final_volume}{volume of the initial dilution that was analyzed for counting each competitor. The units on initial_volume and final_volume must be the same. Example: 80  (If 80 µL were plated.)}
}
}
}
\usage{
competition_template
}
\description{
Data frame template for entering a user's cell count data and
competition parameters for determining relative fitness using the
calculate_fitness() function.
}
\details{
NOTE: If you used the same dilution procedure and analyzed the same volume
of sample at the initial and final timepoints then you can leave the
initial_dilution, initial_volume, final_dilution, and final_volume
columns as the DEFAULT values of 1 or delete them.
}
\examples{
# Write the template to a file that you can edit and use as input to calculate_fitness()
write.csv(competition_template, "my_competition_data.csv")

}
\keyword{datasets}
