% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LTEE_50K_time_course_competitions}
\alias{LTEE_50K_time_course_competitions}
\title{LTEE 50,000 generation time course competition data}
\format{
\subsection{\code{LTEE_50K_time_course_competitions}}{

A data frame with 922 rows and 13 columns:
\describe{
\item{competitor1}{Designation for ancestral strain}
\item{competitor2}{Designation for evolved population}
\item{other columns}{Competition data. See competition_template()}
...
}
}
}
\source{
\url{https://doi.org/10.1126/science.1243357}

\url{https://doi.org/10.5061/dryad.0hc2m}
}
\usage{
LTEE_50K_time_course_competitions
}
\description{
Data showing the fitness effects of adding the rnk-citT module to the
genomes of Escherichia coli strains from the Long-Term Evolution Experiment.
}
\details{
Reference

Wiser M.J., Ribeck N., Lenski R.E. (2013) Long-term dynamics of adaptation
in asexual populations. \emph{Science} 342:1364–1367.
}
\examples{
calculate_fitness(LTEE_50K_time_course_competitions)

}
\keyword{datasets}
