% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{calculate_fitness}
\alias{calculate_fitness}
\title{Calculates relative fitness given a data frame of cell/colony counts}
\usage{
calculate_fitness(input_df, silent = TRUE)
}
\arguments{
\item{input_df}{Input data frame. Must have these columns:
"strain1_initial_count", "strain1_final_count", "strain2_initial_count", "strain2_final_count".}
}
\value{
Data frame with calculated columns added
}
\description{
Calculates relative fitness given a data frame of cell/colony counts
}
\examples{
calculate_fitness(citT_competitions)

}
